requirements.txt
opencv-python==4.5.3.56
numpy==1.21.0
matplotlib==3.4.2
Pillow==8.3.1
